requirements.txt
huggingface_hub==0.22.2
streamlit
requests==2.32.3
pandas==2.1.4
numpy==1.26.4
matplotlib==3.7.1
seaborn==0.13.1
scikit-learn==1.3.2
imbalanced-learn==0.12.3
xgboost==2.1.1
opencv-python==4.10.0.82
Pillow==9.4.0
tensorflow==2.17.0